requirements.txt
# Framework Web
Flask==3.0.0
Flask-WTF==1.2.1
Flask-Cors==4.0.0

# Banco de Dados
supabase==1.0.3
httpx==0.23.3

# Segurança
bcrypt==4.1.1
PyJWT==2.8.0
python-dotenv==1.0.0

# Utilitários
Pillow==10.1.0
python-dateutil==2.8.2
requests==2.31.0
gunicorn==21.2.0

# Desenvolvimento
pytest==7.4.3
black==23.11.0
flake8==6.1.0
mypy==1.7.1